Strip out error_code enum reflection and exceptions.
We only ever expose Exceptions to the user, so there is currently
no need for error_code.

Replace our custom exceptions with sensible built-in exceptions
the user is likely to expect.

In the future, we should probably raise JSONDecodeError for some of
these cases, but simdjson currently can't report pos/lineno/colno.